+++ server/server.js
@@ -1,8 +1,4 @@
-import dotenv from "dotenv";
 import app from "./app.js";
-
-// Load environment variables
-dotenv.config();
 
 // Set port
 const PORT = process.env.PORT || 5000;